Week5-LLaVA/requirements.txt
# Ollama Python client
ollama>=0.1.0

# Image processing
pillow>=10.4.0

# HTTP requests for downloading images from URLs
requests>=2.32.0

# Computer vision for live camera (script 3)
opencv-python>=4.8.0

# Numerical operations
numpy>=1.26.0

